ffff

+++ my_malloc/headers/my_outputs.h
@@ -6,7 +6,7 @@ void 	my_putchar(char c);
 void	my_putstr(char *str);
 void 	my_putnbr(int nbr);
 void 	my_putptr(unsigned long nb);
-int	my_put_hex_nbr(unsigned int nb);
+int	my_put_hex_nbr(unsigned long nb);
 int	my_put_neg(int nb);
 void	my_put_hex(char c);
